+++ src/main/client/client/GUI/Controls/ComplaintCell.java
@@ -54,13 +54,7 @@ public class ComplaintCell extends ListCell<Complaint>{
         {
             if (item != null)
             {
-                lblText.setText("Complaint No. " + item.getComplaintID()
-                        + "Regarding Order No. " + item.getRelatedOrderID()
-                        + "\nComplaint Status: " + item.getStatus()
-                        + (item.getStatus().equals(Complaint.ComplaintStatus.ACCEPTED) ? "\nRefund given: " + item.getRefund() + " NIS (The cheque is in the mail)." : "")
-                        //+"\nAssigned Representative: " + item.getCustomerServiceRepresentative().getName());  // TODO : not working with Employee object. will not compile like this
-                        +"\nAssigned Representative: " + item.getCustomerServiceRepresentativeID());  // workAround -> gets the ID [OrB -> don't want to mess up the gui]
-
+                lblText.setText(item.getGUIString());
 
                 btnDelete.setVisible(item.getStatus().equals(Complaint.ComplaintStatus.NEW) || item.getStatus().equals(Complaint.ComplaintStatus.OPEN));