+++ src/main/client/client/GUI/Forms/Customers/EnterParking.java
@@ -72,7 +72,7 @@ public class EnterParking implements Initializable {
             @Override
             public void run() {
                 Inits.initCars(cmbCar);
-                Inits.initOrders(cmbOrder);
+                Inits.initPreorders(cmbOrder);
             }
         });
         cmbOrder.valueProperty().addListener((observable, oldValue, newValue) -> fillOrder());